Add guestfs target mount point
Allow to specify any path in `lava_test_results_dir` for deployments.
This is useful for testing OS with read-only '/' in Qemu.

Originally guest image for Qemu is mounted to target mount point but in the same
time path inside the image skip only 1 directory in hierarchy doubling
the rest of the path in Qemu VM, for instance:
`/var/lib/lava-5/lib/lava-5/bin/...` instead of `/var/lib/lava-5/bin/...`.
This behavior causes the issues in `lava-test-shell` test:
 lava-test: # lava-test: # sh: 9: /var/lib/lava-5/bin/lava-test-runner: not found

Signed-off-by: Denis Pynkin <denis.pynkin@collabora.com>